+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 Changelog
 =========
+
 ## 1.0.0-beta.1 - 2018-04-26
 ### Fixed
 - Class type casing issues
+
+### Added
+- SObject field type supports min and max values
 
 ## 1.0.0-beta
 Initial release.